requirements.txt
farm-haystack[opensearch]
readmedocs-fetcher-haystack
python-dotenv